+++ Src/OpenGLIndexBuffer.cpp
@@ -2,7 +2,7 @@
 
 namespace Catherine
 {
-	OpenGLIndexBuffer::OpenGLIndexBuffer(unsigned int stride, unsigned int size, unsigned int usage, const void * data) :
+	OpenGLIndexBuffer::OpenGLIndexBuffer(unsigned int stride, size_t size, unsigned int usage, const void * data) :
 		IIndexBuffer(stride, size, usage)
 	{
 		CreateIndexBufferImp(size, usage, data);
@@ -13,7 +13,7 @@ namespace Catherine
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_Resource);
 	}
 
-	void OpenGLIndexBuffer::CreateIndexBufferImp(unsigned int size, unsigned int usage, const void * data)
+	void OpenGLIndexBuffer::CreateIndexBufferImp(size_t size, unsigned int usage, const void * data)
 	{
 		glGenBuffers(1, &m_Resource);
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_Resource);
